Move data set downloading to separate module
